changed {{{em_iter}}} from {{{size_t}}} to {{{unsinged long}}} rather than casting it everywhere, commented type punning fix and vacuous checks in {{{GMTK_FNGramCPT.cc}}}; refs #90 and #98

+++ mitk/bivariate_mi.cc
@@ -1578,7 +1578,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
     time_t timeOfLastSaveParams = time(0);
     time_t timeOfLastPrintNumActive = time(0) - MINTIMEPERPRINTNUMACTIVE -1;
 
-    size_t em_iter=0;
+    unsigned long em_iter=0;
     size_t n_obs_frames;
     size_t * const counts = new size_t [cfr_rng.length()];
      
@@ -1589,7 +1589,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
       em_iter ++;
 
       if ((time(0)-timeOfLastPrint) > NUM_SECONDS_PER_PRINT) {
-	printf("Iter %lu: Starting Iter\n",(unsigned long)em_iter);
+	printf("Iter %lu: Starting Iter\n",em_iter);
 	fflush(stdout);
 	timeOfLastPrint = time(0);
       }
@@ -1610,7 +1610,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
 	//	globalLabelMatrix.printSegmentInfo();
 
 	if ((time(0)-timeOfLastPrint) > NUM_SECONDS_PER_PRINT) {
-	  printf("Iter %lu, sentence %d\n",(unsigned long)em_iter,(*srit));
+	  printf("Iter %lu, sentence %d\n",em_iter,(*srit));
 	    fflush(stdout);
 	    timeOfLastPrint = time(0);
 	}
@@ -1792,7 +1792,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
 
       mbn.endEpoch();
       if ((time(0)-timeOfLastPrint) > NUM_SECONDS_PER_PRINT) {
-	printf("Iter %lu: Finished Iter\n",(unsigned long)em_iter);
+	printf("Iter %lu: Finished Iter\n",em_iter);
 	fflush(stdout);
 	timeOfLastPrint = time(0);
       }
@@ -1807,7 +1807,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
       if ((numActive < prevNumActive) || 
 	  ((time(0) - timeOfLastPrintNumActive) > MINTIMEPERPRINTNUMACTIVE)) {
 	printf("Iter %lu: NA=%d/%lu(%.0f%%), PNA=%d, dist Max(%e) Avg(%e) Min(%e)\n",
-	       (unsigned long)em_iter,
+	       em_iter,
 	       numActive,(unsigned long)n_mis,100*numActive/(double)n_mis,
 	       prevNumActive,
 	       max_dist,
@@ -1824,7 +1824,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
         // the middle
 	// of saving the parameters. So, instead of saving, we forfeit
 	// the work done during this em_iter for safety's sake.
-	printf("Iter %lu: Not Saving Mixture Parameters Since Received SIGUSR2.\n",(unsigned long)em_iter); fflush(stdout);
+	printf("Iter %lu: Not Saving Mixture Parameters Since Received SIGUSR2.\n",em_iter); fflush(stdout);
       } else if (po_fp != NULL && 
 		 ((numActive+nacps <= prevNumActive) ||
 		  (numActive == num_active_to_stop) ||
@@ -1839,7 +1839,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
     delete counts;
 
     if (usr2_terminate) {
-      printf("Iter %lu: Exiting early with failure due to received SIGUSR2\n",(unsigned long)em_iter);
+      printf("Iter %lu: Exiting early with failure due to received SIGUSR2\n",em_iter);
       exit (EXIT_FAILURE);
     }
 
@@ -1854,7 +1854,7 @@ pfile_mi_mg(FILE *out_fp, // where to put output MI values
       fflush(stdout);
     } else {
       printf("Early stop at EM iter %lu, Num Active=%d, not computing MI\n",
-	     (unsigned long)em_iter,numActive);
+	     em_iter,numActive);
     }
 
     // restore signals.
